requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==2.0.0
MarkupSafe==1.1.1
python-digitalocean==1.16.0
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
